OrderFacade methods used only by frontend API has been moved to frontend API
