+++ packages/framework/src/Model/Order/OrderRepository.php
@@ -94,34 +94,6 @@ class OrderRepository
     }
 
     /**
-     * @param string $uuid
-     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
-     * @return \Shopsys\FrameworkBundle\Model\Order\Order|null
-     */
-    protected function findByUuidAndCustomerUser(string $uuid, CustomerUser $customerUser)
-    {
-        return $this->createOrderQueryBuilder()
-            ->andWhere('o.uuid = :uuid')->setParameter(':uuid', $uuid)
-            ->andWhere('o.customerUser = :customerUser')->setParameter(':customerUser', $customerUser)
-            ->setMaxResults(1)
-            ->getQuery()->getOneOrNullResult();
-    }
-
-    /**
-     * @param string $uuid
-     * @param string $urlHash
-     * @return \Shopsys\FrameworkBundle\Model\Order\Order|null
-     */
-    protected function findByUuidAndUrlHash(string $uuid, string $urlHash)
-    {
-        return $this->createOrderQueryBuilder()
-            ->andWhere('o.uuid = :uuid')->setParameter(':uuid', $uuid)
-            ->andWhere('o.urlHash = :urlHash')->setParameter(':urlHash', $urlHash)
-            ->setMaxResults(1)
-            ->getQuery()->getOneOrNullResult();
-    }
-
-    /**
      * @param int $id
      * @return \Shopsys\FrameworkBundle\Model\Order\Order
      */
@@ -131,44 +103,6 @@ class OrderRepository
 
         if ($order === null) {
             throw new \Shopsys\FrameworkBundle\Model\Order\Exception\OrderNotFoundException('Order with ID ' . $id . ' not found.');
-        }
-
-        return $order;
-    }
-
-    /**
-     * @param string $uuid
-     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
-     * @return \Shopsys\FrameworkBundle\Model\Order\Order
-     */
-    public function getByUuidAndCustomerUser(string $uuid, CustomerUser $customerUser): Order
-    {
-        $order = $this->findByUuidAndCustomerUser($uuid, $customerUser);
-
-        if ($order === null) {
-            throw new \Shopsys\FrameworkBundle\Model\Order\Exception\OrderNotFoundException(sprintf(
-                'Order with UUID \'%s\' not found.',
-                $uuid
-            ));
-        }
-
-        return $order;
-    }
-
-    /**
-     * @param string $uuid
-     * @param string $urlHash
-     * @return \Shopsys\FrameworkBundle\Model\Order\Order
-     */
-    public function getByUuidAndUrlHash(string $uuid, string $urlHash): Order
-    {
-        $order = $this->findByUuidAndUrlHash($uuid, $urlHash);
-
-        if ($order === null) {
-            throw new \Shopsys\FrameworkBundle\Model\Order\Exception\OrderNotFoundException(sprintf(
-                'Order with UUID \'%s\' not found.',
-                $uuid
-            ));
         }
 
         return $order;
@@ -241,40 +175,6 @@ class OrderRepository
             ->orderBy('o.createdAt', 'DESC')
             ->setParameter('customerUser', $customerUser)
             ->getQuery()->execute();
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
-     * @param int $limit
-     * @param int $offset
-     * @return \Shopsys\FrameworkBundle\Model\Order\Order[]
-     */
-    public function getCustomerUserOrderLimitedList(CustomerUser $customerUser, int $limit, int $offset): array
-    {
-        return $this->createOrderQueryBuilder()
-            ->andWhere('o.customerUser = :customerUser')
-            ->setParameter('customerUser', $customerUser)
-            ->orderBy('o.createdAt', 'DESC')
-            ->setFirstResult($offset)
-            ->setMaxResults($limit)
-            ->getQuery()
-            ->execute();
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
-     * @return int
-     */
-    public function getCustomerUserOrderCount(CustomerUser $customerUser): int
-    {
-        return $this->em->createQueryBuilder()
-            ->select('count(o.id)')
-            ->from(Order::class, 'o')
-            ->where('o.deleted = FALSE')
-            ->andWhere('o.customerUser = :customerUser')
-            ->setParameter('customerUser', $customerUser)
-            ->getQuery()
-            ->getSingleScalarResult();
     }
 
     /**
