+++ packages/frontend-api/src/Model/Order/OrderFacade.php
@@ -1,0 +1,57 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\FrontendApiBundle\Model\Order;
+
+use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser;
+use Shopsys\FrameworkBundle\Model\Order\Order;
+
+class OrderFacade
+{
+    /**
+     * @var \Shopsys\FrontendApiBundle\Model\Order\OrderRepository
+     */
+    protected $orderRepository;
+
+    /**
+     * @param \Shopsys\FrontendApiBundle\Model\Order\OrderRepository $orderRepository
+     */
+    public function __construct(OrderRepository $orderRepository)
+    {
+        $this->orderRepository = $orderRepository;
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
+     * @param int $limit
+     * @param int $offset
+     * @return \Shopsys\FrameworkBundle\Model\Order\Order[]
+     */
+    public function getCustomerUserOrderLimitedList(
+        CustomerUser $customerUser,
+        int $limit,
+        int $offset
+    ): array {
+        return $this->orderRepository->getCustomerUserOrderLimitedList($customerUser, $limit, $offset);
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
+     * @return int
+     */
+    public function getCustomerUserOrderCount(CustomerUser $customerUser): int
+    {
+        return $this->orderRepository->getCustomerUserOrderCount($customerUser);
+    }
+
+    /**
+     * @param string $uuid
+     * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
+     * @return \Shopsys\FrameworkBundle\Model\Order\Order
+     */
+    public function getByUuidAndCustomerUser(string $uuid, CustomerUser $customerUser): Order
+    {
+        return $this->orderRepository->getByUuidAndCustomerUser($uuid, $customerUser);
+    }
+}
